Added MessageHash to \SpamProtection\Objects

+++ database/message_logs.sql
@@ -9,10 +9,10 @@ create table if not exists spam_protection.message_logs
     user                    blob         null comment 'ZiProto encoded user data',
     forward_from            blob         null comment 'Optional. For forwarded messages, sender of the original message (User ZiProto Blob)',
     forward_from_chat       blob         null comment 'Optional. For messages forwarded from channels, information about the original channel (Chat ZiProto Blob)',
-    forward_from_message_id int          null comment 'Optional. For messages forwarded from channels, identifier of the original message in the channel',
+    forward_from_message_id int(255)     null comment 'Optional. For messages forwarded from channels, identifier of the original message in the channel',
     content_hash            varchar(256) null comment 'The hash of the message content (either from Text or Caption)',
     spam_prediction         float        null comment 'The prediction value for the spam content',
-    ham_prediction          int(255)     null comment 'The prediction value for the ham content',
+    ham_prediction          float        null comment 'The prediction value for the ham content',
     timestamp               int(255)     null comment 'The Unix Timestamp for when this record was created',
     constraint message_logs_id_uindex unique (id),
     constraint message_logs_message_hash_uindex unique (message_hash)
